yanwen: [user-demo] user demo -> main train; main train has the logic of training & evaluating; add ner_evaluator;
